# customgpt-actions-openapi.json
{"openapi":"3.1.0","info":{"title":"Ralph CustomGPT Actions","description":"Lightweight API surface for integrating Ralph with GPT Actions.","version":"0.1.0"},"paths":{"/healthz":{"get":{"tags":["system"],"summary":"Health Check","description":"Basic health probe.","operationId":"health_check_healthz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/runs":{"get":{"tags":["actions"],"summary":"List Action Runs","description":"List all tracked runs (newest first).","operationId":"list_action_runs_runs_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActionListResponse"}}}}},"security":[{"APIKeyHeader":[]}]},"post":{"tags":["actions"],"summary":"Start Action Run","description":"Start an orchestration run and return a polling URL.","operationId":"start_action_run_runs_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActionRunRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActionStartResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/runs/{run_id}":{"get":{"tags":["actions"],"summary":"Get Action Run","description":"Return current status for a run.","operationId":"get_action_run_runs__run_id__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActionRunStatus"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["actions"],"summary":"Cancel Action Run","description":"Cancel a running orchestration.","operationId":"cancel_action_run_runs__run_id__delete","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActionRunStatus"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ActionListResponse":{"properties":{"runs":{"items":{"$ref":"#/components/schemas/ActionRunStatus"},"type":"array","title":"Runs"}},"type":"object","required":["runs"],"title":"ActionListResponse","description":"Response for listing runs."},"ActionRunRequest":{"properties":{"prompt_file":{"type":"string","title":"Prompt File","description":"Path to an existing prompt file"},"config":{"$ref":"#/components/schemas/ConfigOverrides"},"metadata":{"additionalProperties":true,"type":"object","title":"Metadata","description":"Opaque metadata forwarded to status responses"}},"type":"object","required":["prompt_file"],"title":"ActionRunRequest","description":"Payload for starting a new orchestration run."},"ActionRunStatus":{"properties":{"run_id":{"type":"string","title":"Run Id"},"state":{"type":"string","title":"State"},"agent":{"type":"string","title":"Agent"},"prompt_file":{"type":"string","title":"Prompt File"},"started_at":{"type":"string","format":"date-time","title":"Started At"},"completed_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completed At"},"progress":{"additionalProperties":true,"type":"object","title":"Progress"},"artifacts":{"items":{"type":"string"},"type":"array","title":"Artifacts"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"},"metadata":{"additionalProperties":true,"type":"object","title":"Metadata"}},"type":"object","required":["run_id","state","agent","prompt_file","started_at"],"title":"ActionRunStatus","description":"Serialized status for a single run."},"ActionStartResponse":{"properties":{"run_id":{"type":"string","title":"Run Id"},"status_url":{"type":"string","title":"Status Url"}},"type":"object","required":["run_id","status_url"],"title":"ActionStartResponse","description":"Response returned when a run is accepted."},"AgentType":{"type":"string","enum":["claude","gemini","ollama","acp","auto"],"title":"AgentType","description":"Supported AI agent types"},"ConfigOverrides":{"properties":{"agent":{"$ref":"#/components/schemas/AgentType","description":"Which agent adapter to use","default":"auto"},"max_iterations":{"anyOf":[{"type":"integer","maximum":100.0,"minimum":1.0},{"type":"null"}],"title":"Max Iterations","description":"Maximum iterations before stopping"},"max_runtime":{"anyOf":[{"type":"integer","maximum":14400.0,"minimum":60.0},{"type":"null"}],"title":"Max Runtime","description":"Maximum runtime in seconds"},"checkpoint_interval":{"anyOf":[{"type":"integer","minimum":1.0},{"type":"null"}],"title":"Checkpoint Interval","description":"How often to checkpoint iterations"},"max_cost":{"anyOf":[{"type":"number","minimum":0.0},{"type":"null"}],"title":"Max Cost","description":"Maximum spend allowed for the run (USD)"},"ollama_model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ollama Model","description":"Optional Ollama model override when agent=ollama"}},"type":"object","title":"ConfigOverrides","description":"Safe subset of RalphConfig overrides allowed over HTTP."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"APIKeyHeader":{"type":"apiKey","in":"header","name":"x-api-key"}}}}